+++ crates/cli/src/commands/plugin/mod.rs
@@ -2,8 +2,10 @@ mod add;
 mod info;
 mod list;
 mod remove;
+mod search;
 
 pub use add::*;
 pub use info::*;
 pub use list::*;
 pub use remove::*;
+pub use search::*;