+++ crates/cli/src/printer.rs
@@ -23,6 +23,10 @@ impl<'std> Printer<'std> {
         self.buffer.flush().unwrap();
     }
 
+    pub fn write(&mut self, data: impl AsRef<str>) {
+        writeln!(&mut self.buffer, "{}", data.as_ref()).unwrap();
+    }
+
     pub fn line(&mut self) {
         writeln!(&mut self.buffer).unwrap();
     }